+++ php/async/binance.php
@@ -905,7 +905,7 @@ class binance extends Exchange {
                 'fetchCurrencies' => true, // this is a private call and it requires API keys
                 // 'fetchTradesMethod' => 'publicGetAggTrades', // publicGetTrades, publicGetHistoricalTrades, eapiPublicGetTrades
                 'defaultTimeInForce' => 'GTC', // 'GTC' = Good To Cancel (default), 'IOC' = Immediate Or Cancel
-                'defaultType' => 'spot', // 'spot', 'future', 'margin', 'delivery'
+                'defaultType' => 'spot', // 'spot', 'future', 'margin', 'delivery', 'option'
                 'defaultSubType' => null, // 'linear', 'inverse'
                 'hasAlreadyAuthenticatedSuccessfully' => false,
                 'warnOnFetchOpenOrdersWithoutSymbol' => true,
@@ -1436,6 +1436,66 @@ class binance extends Exchange {
         $this->options['sandboxMode'] = $enable;
     }
 
+    public function create_expired_option_market($symbol) {
+        // support expired option contracts
+        $settle = 'USDT';
+        $optionParts = explode('-', $symbol);
+        $symbolBase = explode('/', $symbol);
+        $base = null;
+        if (mb_strpos($symbol, '/') > -1) {
+            $base = $this->safe_string($symbolBase, 0);
+        } else {
+            $base = $this->safe_string($optionParts, 0);
+        }
+        $expiry = $this->safe_string($optionParts, 1);
+        $strike = $this->safe_string($optionParts, 2);
+        $optionType = $this->safe_string($optionParts, 3);
+        return array(
+            'id' => $base . '-' . $expiry . '-' . $strike . '-' . $optionType,
+            'symbol' => $base . '/' . $settle . ':' . $settle . '-' . $expiry . '-' . $strike . '-' . $optionType,
+            'base' => $base,
+            'quote' => $settle,
+            'baseId' => $base,
+            'quoteId' => $settle,
+            'active' => null,
+            'type' => 'option',
+            'linear' => null,
+            'inverse' => null,
+            'spot' => false,
+            'swap' => false,
+            'future' => false,
+            'option' => true,
+            'margin' => false,
+            'contract' => true,
+            'contractSize' => null,
+            'expiry' => null,
+            'expiryDatetime' => null,
+            'optionType' => ($optionType === 'C') ? 'call' : 'put',
+            'strike' => $strike,
+            'settle' => $settle,
+            'settleId' => $settle,
+            'precision' => array(
+                'amount' => null,
+                'price' => null,
+            ),
+            'limits' => array(
+                'amount' => array(
+                    'min' => null,
+                    'max' => null,
+                ),
+                'price' => array(
+                    'min' => null,
+                    'max' => null,
+                ),
+                'cost' => array(
+                    'min' => null,
+                    'max' => null,
+                ),
+            ),
+            'info' => null,
+        );
+    }
+
     public function market($symbol) {
         if ($this->markets === null) {
             throw new ExchangeError($this->id . ' $markets not loaded');
@@ -1486,6 +1546,8 @@ class binance extends Exchange {
                 if (is_array($this->markets) && array_key_exists($futuresSymbol, $this->markets)) {
                     return $this->markets[$futuresSymbol];
                 }
+            } elseif ((mb_strpos($symbol, '-C') > -1) || (mb_strpos($symbol, '-P') > -1)) { // both exchange-id and unified symbols are supported this way regardless of the  $defaultType
+                return $this->create_expired_option_market($symbol);
             }
         }
         throw new BadSymbol($this->id . ' does not have $market $symbol ' . $symbol);
@@ -2550,7 +2612,9 @@ class binance extends Exchange {
                 $request['limit'] = $limit; // default 100, max 5000, see https://github.com/binance/binance-spot-api-docs/blob/master/rest-api.md#order-book
             }
             $method = 'publicGetDepth';
-            if ($market['linear']) {
+            if ($market['option']) {
+                $method = 'eapiPublicGetDepth';
+            } elseif ($market['linear']) {
                 $method = 'fapiPublicGetDepth';
             } elseif ($market['inverse']) {
                 $method = 'dapiPublicGetDepth';
@@ -2574,9 +2638,27 @@ class binance extends Exchange {
             //             ["2493.71","12.054"],
             //         ]
             //     }
+            //
+            // options (eapi)
+            //
+            //     {
+            //         "bids" => [
+            //             ["108.7","16.08"],
+            //             ["106","21.29"],
+            //             ["82.4","0.02"]
+            //         ],
+            //         "asks" => [
+            //             ["111.4","19.52"],
+            //             ["119.9","17.6"],
+            //             ["141.2","31"]
+            //         ],
+            //         "T" => 1676771382078,
+            //         "u" => 1015939
+            //     }
+            //
             $timestamp = $this->safe_integer($response, 'T');
             $orderbook = $this->parse_order_book($response, $symbol, $timestamp);
-            $orderbook['nonce'] = $this->safe_integer($response, 'lastUpdateId');
+            $orderbook['nonce'] = $this->safe_integer_2($response, 'lastUpdateId', 'u');
             return $orderbook;
         }) ();
     }
@@ -2947,13 +3029,30 @@ class binance extends Exchange {
         //         )
         //     )
         //
+        // options
+        //
+        //     {
+        //         "open" => "32.2",
+        //         "high" => "32.2",
+        //         "low" => "32.2",
+        //         "close" => "32.2",
+        //         "volume" => "0",
+        //         "interval" => "5m",
+        //         "tradeCount" => 0,
+        //         "takerVolume" => "0",
+        //         "takerAmount" => "0",
+        //         "amount" => "0",
+        //         "openTime" => 1677096900000,
+        //         "closeTime" => 1677097200000
+        //     }
+        //
         return array(
-            $this->safe_integer($ohlcv, 0),
-            $this->safe_number($ohlcv, 1),
-            $this->safe_number($ohlcv, 2),
-            $this->safe_number($ohlcv, 3),
-            $this->safe_number($ohlcv, 4),
-            $this->safe_number($ohlcv, 5),
+            $this->safe_integer_2($ohlcv, 0, 'closeTime'),
+            $this->safe_number_2($ohlcv, 1, 'open'),
+            $this->safe_number_2($ohlcv, 2, 'high'),
+            $this->safe_number_2($ohlcv, 3, 'low'),
+            $this->safe_number_2($ohlcv, 4, 'close'),
+            $this->safe_number_2($ohlcv, 5, 'volume'),
         );
     }
 
@@ -3009,7 +3108,9 @@ class binance extends Exchange {
                 $request['endTime'] = $until;
             }
             $method = 'publicGetKlines';
-            if ($price === 'mark') {
+            if ($market['option']) {
+                $method = 'eapiPublicGetKlines';
+            } elseif ($price === 'mark') {
                 if ($market['inverse']) {
                     $method = 'dapiPublicGetMarkPriceKlines';
                 } else {
@@ -3033,6 +3134,25 @@ class binance extends Exchange {
             //         [1591478580000,"0.02499600","0.02500900","0.02499400","0.02500300","21.34700000",1591478639999,"0.53370468",24,"7.53800000","0.18850725","0"],
             //         [1591478640000,"0.02500800","0.02501100","0.02500300","0.02500800","154.14200000",1591478699999,"3.85405839",97,"5.32300000","0.13312641","0"],
             //     ]
+            //
+            // options (eapi)
+            //
+            //     array(
+            //         {
+            //             "open" => "32.2",
+            //             "high" => "32.2",
+            //             "low" => "32.2",
+            //             "close" => "32.2",
+            //             "volume" => "0",
+            //             "interval" => "5m",
+            //             "tradeCount" => 0,
+            //             "takerVolume" => "0",
+            //             "takerAmount" => "0",
+            //             "amount" => "0",
+            //             "openTime" => 1677096900000,
+            //             "closeTime" => 1677097200000
+            //         }
+            //     )
             //
             return $this->parse_ohlcvs($response, $market, $timeframe, $since, $limit);
         }) ();
@@ -3175,7 +3295,8 @@ class binance extends Exchange {
         $amount = $this->safe_string($trade, 'quantity', $amount);
         $cost = $this->safe_string_2($trade, 'quoteQty', 'baseQty');  // inverse futures
         $marketId = $this->safe_string($trade, 'symbol');
-        $marketType = (is_array($trade) && array_key_exists('isIsolated', $trade)) || (is_array($trade) && array_key_exists('M', $trade)) || (is_array($trade) && array_key_exists('orderListId', $trade)) ? 'spot' : 'contract';
+        $isSpotTrade = (is_array($trade) && array_key_exists('isIsolated', $trade)) || (is_array($trade) && array_key_exists('M', $trade)) || (is_array($trade) && array_key_exists('orderListId', $trade));
+        $marketType = $isSpotTrade ? 'spot' : 'contract';
         $market = $this->safe_market($marketId, $market, null, $marketType);
         $symbol = $market['symbol'];
         $id = $this->safe_string_2($trade, 't', 'a');
@@ -3208,11 +3329,6 @@ class binance extends Exchange {
         }
         if ((is_array($trade) && array_key_exists('optionSide', $trade)) || $market['option']) {
             $settle = $this->safe_currency_code($this->safe_string($trade, 'quoteAsset', 'USDT'));
-            if ($symbol === null) {
-                $optionParts = explode('-', $marketId);
-                $optionType = $this->safe_string($optionParts, 3);
-                $symbol = $this->safe_string($optionParts, 0) . '/' . $settle . ':' . $settle . '-' . $this->safe_string($optionParts, 1) . '-' . $this->safe_string($optionParts, 2) . '-' . $optionType;
-            }
             $takerOrMaker = $this->safe_string_lower($trade, 'liquidity');
             if (is_array($trade) && array_key_exists('fee', $trade)) {
                 $fee = array(
@@ -4351,31 +4467,35 @@ class binance extends Exchange {
              * @param {array} $params extra parameters specific to the binance api endpoint
              * @return {[array]} a list of {@link https://docs.ccxt.com/en/latest/manual.html#trade-structure trade structures}
              */
-            if ($symbol === null) {
-                throw new ArgumentsRequired($this->id . ' fetchMyTrades() requires a $symbol argument');
-            }
             Async\await($this->load_markets());
-            $market = $this->market($symbol);
-            $request = array(
-                'symbol' => $market['id'],
-            );
-            $type = $this->safe_string($params, 'type', $market['type']);
-            $params = $this->omit($params, 'type');
+            $request = array();
+            $market = null;
+            $type = null;
             $method = null;
             $marginMode = null;
-            list($marginMode, $params) = $this->handle_margin_mode_and_params('fetchMyTrades', $params);
-            if ($type === 'spot' || $type === 'margin') {
-                $method = 'privateGetMyTrades';
-                if (($type === 'margin') || ($marginMode !== null)) {
-                    $method = 'sapiGetMarginMyTrades';
-                    if ($marginMode === 'isolated') {
-                        $request['isIsolated'] = true;
+            if ($symbol !== null) {
+                $market = $this->market($symbol);
+                $request['symbol'] = $market['id'];
+            }
+            list($type, $params) = $this->handle_market_type_and_params('fetchMyTrades', $market, $params);
+            if ($type === 'option') {
+                $method = 'eapiPrivateGetUserTrades';
+            } else {
+                $this->check_required_symbol('fetchMyTrades', $symbol);
+                list($marginMode, $params) = $this->handle_margin_mode_and_params('fetchMyTrades', $params);
+                if ($type === 'spot' || $type === 'margin') {
+                    $method = 'privateGetMyTrades';
+                    if (($type === 'margin') || ($marginMode !== null)) {
+                        $method = 'sapiGetMarginMyTrades';
+                        if ($marginMode === 'isolated') {
+                            $request['isIsolated'] = true;
+                        }
                     }
+                } elseif ($market['linear']) {
+                    $method = 'fapiPrivateGetUserTrades';
+                } elseif ($market['inverse']) {
+                    $method = 'dapiPrivateGetUserTrades';
                 }
-            } elseif ($market['linear']) {
-                $method = 'fapiPrivateGetUserTrades';
-            } elseif ($market['inverse']) {
-                $method = 'dapiPrivateGetUserTrades';
             }
             $endTime = $this->safe_integer_2($params, 'until', 'endTime');
             if ($since !== null) {
@@ -4399,7 +4519,7 @@ class binance extends Exchange {
                 $params = $this->omit($params, array( 'endTime', 'until' ));
             }
             if ($limit !== null) {
-                if ($market['contract']) {
+                if (($type === 'option') || $market['contract']) {
                     $limit = min ($limit, 1000); // above 1000, returns error
                 }
                 $request['limit'] = $limit;
@@ -4443,6 +4563,30 @@ class binance extends Exchange {
             //             "side" => "SELL",
             //             "symbol" => "BTCUSDT",
             //             "time" => 1569514978020
+            //         }
+            //     )
+            //
+            // options (eapi)
+            //
+            //     array(
+            //         {
+            //             "id" => 1125899906844226012,
+            //             "tradeId" => 73,
+            //             "orderId" => 4638761100843040768,
+            //             "symbol" => "ETH-230211-1500-C",
+            //             "price" => "18.70000000",
+            //             "quantity" => "-0.57000000",
+            //             "fee" => "0.17305890",
+            //             "realizedProfit" => "-3.53400000",
+            //             "side" => "SELL",
+            //             "type" => "LIMIT",
+            //             "volatility" => "0.30000000",
+            //             "liquidity" => "MAKER",
+            //             "time" => 1676085216845,
+            //             "priceScale" => 1,
+            //             "quantityScale" => 2,
+            //             "optionSide" => "CALL",
+            //             "quoteAsset" => "USDT"
             //         }
             //     )
             //