+++ php/btcex.php
@@ -1593,19 +1593,18 @@ class btcex extends Exchange {
             // 'end_id' => 0, // The ID number of the last trade to be returned
             // 'sorting' => '', // Direction of results sorting,default => desc
             // 'self_trade' => false, // If not set, query all
+            // 'start_timestamp' => false // The trade time of the first trade to be returned
+            // 'end_timestamp' => false // The trade time of the last trade to be returned
         );
-        $method = null;
         $market = $this->market($symbol);
         $request['instrument_name'] = $market['id'];
-        if ($since === null) {
-            $method = 'privateGetGetUserTradesByInstrument';
-        } else {
-            $method = 'privateGetGetUserTradesByInstrumentAndTime';
-        }
         if ($limit !== null) {
             $request['count'] = $limit; // default 20
         }
-        $response = $this->$method (array_merge($request, $params));
+        if ($since !== null) {
+            $request['start_timestamp'] = $since;
+        }
+        $response = $this->privateGetGetUserTradesByInstrument (array_merge($request, $params));
         $result = $this->safe_value($response, 'result', array());
         //
         //     {